MAINT: Migrate nose to pytest
