+++ seaborn/tests/test_rcmod.py
@@ -1,16 +1,16 @@
 from distutils.version import LooseVersion
 
+import pytest
 import numpy as np
 import matplotlib as mpl
-import nose
 import matplotlib.pyplot as plt
-import nose.tools as nt
 import numpy.testing as npt
 
 from .. import rcmod, palettes, utils
+from ..conftest import has_verdana
 
 
-class RCParamTester(object):
+class RCParamTester:
 
     def flatten_list(self, orig_list):
 
@@ -28,7 +28,7 @@ class RCParamTester(object):
             if isinstance(v, np.ndarray):
                 npt.assert_array_equal(mpl.rcParams[k], v)
             else:
-                nt.assert_equal((k, mpl.rcParams[k]), (k, v))
+                assert mpl.rcParams[k] == v
 
     def assert_rc_params_equal(self, params1, params2):
 
@@ -42,7 +42,7 @@ class RCParamTester(object):
             if isinstance(v1, np.ndarray):
                 npt.assert_array_equal(v1, v2)
             else:
-                nt.assert_equal(v1, v2)
+                assert v1 == v2
 
 
 class TestAxesStyle(RCParamTester):
@@ -58,19 +58,19 @@ class TestAxesStyle(RCParamTester):
 
         _style_keys = set(rcmod._style_keys)
         for style in self.styles:
-            nt.assert_true(not set(rcmod.axes_style(style)) ^ _style_keys)
+            assert not set(rcmod.axes_style(style)) ^ _style_keys
 
     def test_bad_style(self):
 
-        with nt.assert_raises(ValueError):
+        with pytest.raises(ValueError):
             rcmod.axes_style("i_am_not_a_style")
 
     def test_rc_override(self):
 
         rc = {"axes.facecolor": "blue", "foo.notaparam": "bar"}
         out = rcmod.axes_style("darkgrid", rc)
-        nt.assert_equal(out["axes.facecolor"], "blue")
-        nt.assert_not_in("foo.notaparam", out)
+        assert out["axes.facecolor"] == "blue"
+        assert "foo.notaparam" not in out
 
     def test_set_style(self):
 
@@ -98,12 +98,12 @@ class TestAxesStyle(RCParamTester):
 
     def test_style_context_independence(self):
 
-        nt.assert_true(set(rcmod._style_keys) ^ set(rcmod._context_keys))
+        assert set(rcmod._style_keys) ^ set(rcmod._context_keys)
 
     def test_set_rc(self):
 
         rcmod.set_theme(rc={"lines.linewidth": 4})
-        nt.assert_equal(mpl.rcParams["lines.linewidth"], 4)
+        assert mpl.rcParams["lines.linewidth"] == 4
         rcmod.set_theme()
 
     def test_set_with_palette(self):
@@ -169,11 +169,11 @@ class TestPlottingContext(RCParamTester):
         _context_keys = set(rcmod._context_keys)
         for context in self.contexts:
             missing = set(rcmod.plotting_context(context)) ^ _context_keys
-            nt.assert_true(not missing)
+            assert not missing
 
     def test_bad_context(self):
 
-        with nt.assert_raises(ValueError):
+        with pytest.raises(ValueError):
             rcmod.plotting_context("i_am_not_a_context")
 
     def test_font_scale(self):
@@ -188,15 +188,15 @@ class TestPlottingContext(RCParamTester):
             font_keys.append("legend.title_fontsize")
 
         for k in font_keys:
-            nt.assert_equal(notebook_ref[k] * 2, notebook_big[k])
+            assert notebook_ref[k] * 2 == notebook_big[k]
 
     def test_rc_override(self):
 
         key, val = "grid.linewidth", 5
         rc = {key: val, "foo": "bar"}
         out = rcmod.plotting_context("talk", rc=rc)
-        nt.assert_equal(out[key], val)
-        nt.assert_not_in("foo", out)
+        assert out[key] == val
+        assert "foo" not in out
 
     def test_set_context(self):
 
@@ -223,7 +223,7 @@ class TestPlottingContext(RCParamTester):
         self.assert_rc_params(orig_params)
 
 
-class TestPalette(object):
+class TestPalette:
 
     def test_set_palette(self):
 
@@ -240,8 +240,9 @@ class TestPalette(object):
         assert utils.get_color_cycle() == palettes.color_palette("Set2", 8)
 
 
-class TestFonts(object):
+class TestFonts:
 
+    @pytest.mark.skipif(not has_verdana(), reason="Verdana font is not present")
     def test_set_font(self):
 
         rcmod.set_theme(font="Verdana")
@@ -249,16 +250,9 @@ class TestFonts(object):
         _, ax = plt.subplots()
         ax.set_xlabel("foo")
 
-        try:
-            nt.assert_equal(ax.xaxis.label.get_fontname(),
-                            "Verdana")
-        except AssertionError:
-            if has_verdana():
-                raise
-            else:
-                raise nose.SkipTest("Verdana font is not present")
-        finally:
-            rcmod.set_theme()
+        assert ax.xaxis.label.get_fontname() == "Verdana"
+
+        rcmod.set_theme()
 
     def test_set_serif_font(self):
 
@@ -267,11 +261,11 @@ class TestFonts(object):
         _, ax = plt.subplots()
         ax.set_xlabel("foo")
 
-        nt.assert_in(ax.xaxis.label.get_fontname(),
-                     mpl.rcParams["font.serif"])
+        assert ax.xaxis.label.get_fontname() in mpl.rcParams["font.serif"]
 
         rcmod.set_theme()
 
+    @pytest.mark.skipif(not has_verdana(), reason="Verdana font is not present")
     def test_different_sans_serif(self):
 
         rcmod.set_theme()
@@ -280,36 +274,6 @@ class TestFonts(object):
         _, ax = plt.subplots()
         ax.set_xlabel("foo")
 
-        try:
-            nt.assert_equal(ax.xaxis.label.get_fontname(),
-                            "Verdana")
-        except AssertionError:
-            if has_verdana():
-                raise
-            else:
-                raise nose.SkipTest("Verdana font is not present")
-        finally:
-            rcmod.set_theme()
+        assert ax.xaxis.label.get_fontname() == "Verdana"
 
-
-def has_verdana():
-    """Helper to verify if Verdana font is present"""
-    # This import is relatively lengthy, so to prevent its import for
-    # testing other tests in this module not requiring this knowledge,
-    # import font_manager here
-    import matplotlib.font_manager as mplfm
-    try:
-        verdana_font = mplfm.findfont('Verdana', fallback_to_default=False)
-    except:  # noqa
-        # if https://github.com/matplotlib/matplotlib/pull/3435
-        # gets accepted
-        return False
-    # otherwise check if not matching the logic for a 'default' one
-    try:
-        unlikely_font = mplfm.findfont("very_unlikely_to_exist1234",
-                                       fallback_to_default=False)
-    except:  # noqa
-        # if matched verdana but not unlikely, Verdana must exist
-        return True
-    # otherwise -- if they match, must be the same default
-    return verdana_font != unlikely_font
+        rcmod.set_theme()
